[CHORE] Create separate package for pr and issue model classes
