+++ src/test/java/com/mcp/github/services/tests/GithubServiceTest.java
@@ -4,9 +4,9 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 
 import com.mcp.github.client.GithubHttpClient;
-import com.mcp.github.models.Issue;
-import com.mcp.github.models.IssueDetail;
-import com.mcp.github.models.PullRequestSummary;
+import com.mcp.github.models.issue.Issue;
+import com.mcp.github.models.issue.IssueDetail;
+import com.mcp.github.models.pr.PullRequestSummary;
 import com.mcp.github.services.GithubService;
 import java.time.Clock;
 import java.time.Instant;
